Setup easy backtracking alt system
Once matching is happening, I can switch in a lockstep algorithm.

I think the way to do that is to have a struct which tracks progress in
each sub-regex, so that I can handle tasks such as groups with an inner
alt set, and still make progress one byte at a time.

The no-allocation-allowed thing means I'll need five of those, for each
of one, two, three, four, and 'more' alts.  That sounds like an recipie
for combinatoric explosion.  Time to revisit all this once we have some
tests.