Miscellaneous yeoman work on character escapes
I plan to circle back and build character masks for all the standard
sets, and include those in escape sequences as well.

That will poke so many holes in the ASCII range that I may as well
make the high 128 into a range and make feeding the character sets the
else branch.