fix: word breaks on zero-length strings
Mvzr did not originally account for the case of matching a zero-length
string, at all.  While it was unclear to me what the semantics of word-
break or not-word-break should even be, PCRE2 thinks neither match an
empty string.  So now `mvzr` doesn't either, and also, doesn't try to
read a character from a string which doesn't have one.